+++ frontend/src/hooks/pages/useTravelPlanFormState/useTravelPlanFormState.ts
@@ -1,6 +1,6 @@
 import { TravelTransformDays } from "@type/domain/travelTransform";
 
-import { useTravelPlanDays } from "@hooks/pages/useTravelPlanDays";
+import { useTravelPlanDays } from "@hooks/pages/useTravelPlanFormState/useTravelPlanDays";
 import useTravelPlanStartDate from "@hooks/pages/useTravelPlanFormState/useTravelPlanStartDate";
 import useTravelPlanTitle from "@hooks/pages/useTravelPlanFormState/useTravelPlanTitle";
