+++ frontend/src/hooks/pages/useTravelogueFormState/useTravelogueThumbnail.ts
